create a subdir for podcasts
If a podcast has episodes, a subdir will be created where all episodes
are saved. The subdir name depends on the podcast parent name and the
`filename_name` mode option.

If downloading a single episode via the `--title` or `--asin` option
(`--resolve-podcasts` flag must be set for this) the episode will be
stored in the dir set by `--output-dir` option!