fix an issue with requesting download urls for aax files (#62)
In some cases, requesting a download url with the new method fails. It
will give a url but with a wrong codec. These url result in an error,
when downloading these files.
As a workaround I've added the old download url creation method. This
will work until I found a better method!