Update CHANGELOG.md and rework plugin module
In this commit, the CHANGELOG.md file was updated with additional plugin information and a change in the plugin module. The plugin command help page now includes information about the source of the plugin and plugin commands are clearly indicated with a `(P)` prefix. Additionally, the plugin module was reworked: now it relies on importlib.metadata instead of setuptools (pkg_resources) to get entry points.