fix: decrypt command filename escape
`subprocess.check_output()` is not called with the option `shell = True` so it doesn't make sense to escape any arguments because ffmpeg will already receive them as is.

When e.g. a filename contains escapable characters, then the `quote()` function encloses the it with quotation marks ''. FFmpeg interprets those literally and they don't match any file (the actual file doesn't start with ').